requirements.txt
fastapi
uvicorn
Pillow
python-dotenv
google-genai
python-multipart
supabase